add crud-processes-of-admin

+++ app/Http/SystemUtils.php
@@ -17,6 +17,7 @@ class SystemUtils{
 
             if(!empty($filename))
             {
+                if($filename!=="default.png")
                 unlink($destination.'/' . $filename);
             }
